escpos_web/requirements.txt
# Only for escpos_web
Django>=3.2,<3.3
djangorestframework>=3.12,<3.13
djangorestframework-filters==1.0.0.dev0
channels>=3.0,<3.1
netifaces>=0.11,<0.12
python-escpos==v3.0a8
websockets>=9.1,<9.2
numpy>=1.16,<1.17
RPi.GPIO>=0.7,<0.8
spidev>=3.5,<3.6